requirements.txt
groq>=0.4.1
python-dotenv>=1.0.0
pymongo>=4.0.0
pydantic>=2.5.0
datetime
typing
bson
